Fix compile size check

+++ src/ray/core_worker/store_provider/plasma_store_provider.cc
@@ -227,7 +227,7 @@ Status CoreWorkerPlasmaStoreProvider::Wait(const std::unordered_set<ObjectID> &o
     RAY_RETURN_NOT_OK(raylet_client_->Wait(id_vector, num_objects, call_timeout, false,
                                            task_id, &result_pair));
 
-    if (result_pair.first.size() >= num_objects) {
+    if (result_pair.first.size() >= static_cast<size_t>(num_objects)) {
       should_break = true;
     }
     if (check_signals_) {
